backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.2
colorama==0.4.6
cryptography==46.0.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
google-auth==2.49.1
idna==3.11
joblib==1.5.3
kombu==5.6.2
llvmlite==0.47.0
numba==0.65.0
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
redis==7.4.0
requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
shap==0.51.0
six==1.17.0
slicer==0.0.8
sqlparse==0.5.5
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
xgboost==3.2.0
